...move page stuff back into primary ingestion pipeline
