+++ services/ingestion/src/pdf_ingestion.py
@@ -11,15 +11,18 @@ import tempfile
 import time
 import os
 import glob
+import subprocess
+from PIL import Image
 from typing import Mapping, TypeVar, Callable
 from pdf_extractor import parse_pdf
 from pymongo import MongoClient
+from bson.objectid import ObjectId
 import json
 
 
 T = TypeVar('T')
 
-def run_pdf_ingestion(pdf_dir: str, db_insert_fn: Callable[[Mapping[T, T]], None]) -> Mapping[T, T]:
+def run_pdf_ingestion(pdf_dir: str, db_insert_fn: Callable[[Mapping[T, T]], None], db_insert_pages_fn: Callable[[Mapping[T,T]], None], subprocess_fn: Callable[[str, str], None]) -> Mapping[T, T]:
     """
     Entry point for ingesting PDF documents
     """
@@ -31,6 +34,13 @@ def run_pdf_ingestion(pdf_dir: str, db_insert_fn: Callable[[Mapping[T, T]], None
         with tempfile.TemporaryDirectory() as img_tmp:
             pdf_obj = {}
             pdf_obj = load_pdf_metadata(pdf_path, pdf_obj)
+            pdf_obj["_id"] = ObjectId() # Want to know the _id _before_ insertion so we can tag pages in their collection
+            pdf_path = os.path.abspath("%s/%s" % (pdf_dir, pdf_obj["pdf_name"]))
+            # TODO: get real pdf_path as pdf_dir + pdf_name (from doc)
+            subprocess_fn(pdf_path, img_tmp)
+            pages = []
+            pages = load_page_data(img_tmp, pdf_obj)
+            db_insert_pages_fn(pages)
             pdfs.append(pdf_obj)
 
     docids = db_insert_fn(pdfs)
@@ -47,7 +57,39 @@ def insert_pdfs_mongo(pdfs: Mapping[T, T]) -> None:
     db = client.pdfs
     pdf_collection = db.raw_pdfs
     result = pdf_collection.insert_many(pdfs)
-    # TODO: page_data should ideally get stored in a page collection with a link up to these docid (?)
+
+def run_ghostscript(pdf_path: str, img_tmp: str) -> None:
+    """
+    Run ghostscript as a subprocess over pdf files
+    """
+    with tempfile.TemporaryFile() as gs_stdout, tempfile.TemporaryFile() as gs_stderr:
+        subprocess.run(['gs', '-dBATCH',
+                              '-dNOPAUSE',
+                              '-sDEVICE=png16m',
+                              '-dGraphicsAlphaBits=4',
+                              '-dTextAlphaBits=4',
+                              '-r600',
+                              f'-sOutputFile="{img_tmp}/%d"',
+                              pdf_path
+                        ], stdout=gs_stdout, stderr=gs_stderr)
+        out = gs_stdout.read()
+        err = gs_stdout.read()
+        if len(out):
+            logging.info("Ghostscript output:")
+            logging.info(str(out))
+        if len(err):
+            logging.warning("Ghostscript err: ")
+            logging.warning(err)
+
+def insert_pages_mongo(pages: Mapping[T, T]) -> None:
+    """
+    Insert pdfs into mongodb
+    TODO: Pass correct config
+    """
+    client = MongoClient(os.environ["DBCONNECT"])
+    db = client.pdfs
+    page_collection = db.pages
+    result = page_collection.insert_many(pages)
 
 
 def load_page_data(img_dir: str, current_obj: Mapping[T, T]) -> Mapping[T, T]:
@@ -67,10 +109,10 @@ def load_page_data(img_dir: str, current_obj: Mapping[T, T]) -> Mapping[T, T]:
         page_obj['page_width'] = width
         page_obj['page_height'] = height
         page_obj['page_num'] = page_num
+        page_obj['pdf_name'] = current_obj['pdf_name']
+        page_obj['pdf_id'] = current_obj['_id']
         page_data.append(page_obj)
-    current_obj['page_data'] = page_data
-    current_obj['event_stream'].append('imagedata')
-    return current_obj
+    return page_data
 
 
 def load_pdf_metadata(pdf_path: str, current_obj: Mapping[T, T]) -> Mapping[T, T]:
@@ -94,7 +136,7 @@ def load_pdf_metadata(pdf_path: str, current_obj: Mapping[T, T]) -> Mapping[T, T
 @click.command()
 @click.argument('pdf_dir')
 def click_wrapper(pdf_dir: str) -> None:
-    run_pdf_ingestion(pdf_dir, insert_pdfs_mongo)
+    run_pdf_ingestion(pdf_dir, insert_pdfs_mongo, insert_pages_mongo, run_ghostscript)
 
 if __name__ == '__main__':
     click_wrapper()
