--- services/page_extraction/install.sh
@@ -1,1 +1,0 @@
-helm install --set mongopwd=$MONGODB_ROOT_PASSWORD .
