--- services/page_extraction/requirements.txt
@@ -1,3 +1,0 @@
-pillow
-click
-pymongo
