--- services/page_extraction/templates/deployment.yaml
@@ -1,37 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: {{ include "page_extraction.fullname" . }}
-  labels:
-    app.kubernetes.io/name: {{ include "page_extraction.name" . }}
-    helm.sh/chart: {{ include "page_extraction.chart" . }}
-    app.kubernetes.io/instance: {{ .Release.Name }}
-    app.kubernetes.io/managed-by: {{ .Release.Service }}
-spec:
-  replicas: 1
-  selector:
-      matchLabels:
-        app.kubernetes.io/name: {{ include "page_extraction.name" . }}
-        app.kubernetes.io/instance: {{ .Release.Name }}
-  template:
-    metadata:
-      labels:
-        app.kubernetes.io/name: {{ include "page_extraction.name" . }}
-        app.kubernetes.io/instance: {{ .Release.Name }}
-    spec:
-      containers:
-        - name: {{ .Chart.Name }}
-          image: "{{ .Values.image.repository }}:{{ .Values.image.tag }}"
-          imagePullPolicy: {{ .Values.image.pullPolicy }}
-          volumeMounts:
-              - mountPath: /input
-                name: input-volume
-          command: ['python3', 'page_extraction.py', '/input']
-          env:
-              - name: DBCONNECT
-                value: "mongodb://{{ .Values.mongouser }}:{{ .Values.mongopwd }}@{{ .Values.mongourl }}:{{ .Values.mongoport }}"
-      restartPolicy: Always
-      volumes:
-          - name: input-volume
-            hostPath:
-                path: {{ .Values.pdfDir }}
