--- services/page_extraction/test_install.sh
@@ -1,1 +1,0 @@
-helm install --dry-run --debug --set mongopwd=$MONGODB_ROOT_PASSWORD .
